requirements.txt
mcp>=1.3.0
qdrant-client>=1.7.0
fastembed>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.27.0
starlette>=0.37.0
backoff>=2.2.1
prometheus-client>=0.19.0
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-docker>=3.1.1
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
ruff>=0.1.0
numpy>=1.24.0
jinja2>=3.1.0
pyyaml>=6.0.0
docker>=7.0.0
fastapi>=0.110.0 